+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-Nothing yet!
+* 新增 `queryParamsUrl()` 函數的 `clear` 參數
 
 ## v3.0.1 - 2025-07-05